BinaryCodes/Web/medium-xss1/website/requirements.txt
Django==1.11
django-csp==3.3
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
MySQL-python==1.2.5
Pillow==5.0.0
pytz==2017.3
selenium==3.8.1
six==1.11.0
gunicorn
requests
